Do not use lirideployment
Qbs 1.10 wants us to specify a priority when there are multiple instances
of modules with the same name.

Unfortunately in a system where qbs-shared is already installed we are
in that situation: one instance from the system and another one from
the source tree.

The solution is to never use our own modules again, we just need
a couple of properties to specify the installation paths and
that's it.